pyproject.toml
[project]
name = "vix-alert-bot"
version = "0.1.0"
description = "VIX mean-reversion trading signal bot"
requires-python = ">=3.11"
dependencies = [
    # Data fetching
    "yfinance>=0.2.36",
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "lxml>=5.1",
    # Data processing
    "pandas>=2.2",
    "numpy>=1.26",
    # ML - Training
    "xgboost>=2.0",
    "scikit-learn>=1.4",
    "torch>=2.2",
    # ONNX export & inference
    "onnx>=1.15",
    "onnxruntime>=1.17",
    "onnxmltools>=1.12",
    "skl2onnx>=1.16",
    # Visualization
    "matplotlib>=3.8",
    "seaborn>=0.13",
    # Bot infrastructure
    "python-telegram-bot>=21.0",
    "apscheduler>=3.10",
    "exchange-calendars>=4.5",
    # IBKR
    "ib_insync>=0.9.86",
    # Config
    "python-dotenv>=1.0",
    # Web (health endpoint)
    "fastapi>=0.109",
    "uvicorn>=0.27",
    # Database
    "aiosqlite>=0.19",
    # Jupyter
    "jupyter>=1.0",
    "ipykernel>=6.29",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.2",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["bot", "training", "data"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
